StrategyRegistry extracted
