+++ WebScrapingWorker/src/Modules/DataGeneration/DataGeneration.ts
@@ -1,6 +1,6 @@
 import { injectable } from 'inversify';
 import { IDataGeneration } from './IDataGeneration.js';
-import { IDataGenerationStrategy, DataGenerationStrategyRegistry } from './IDataGenerationStrategy.js';
+import { IDataGenerationStrategy, dataGenerationStrategyRegistry } from './IDataGenerationStrategy.js';
 
 @injectable()
 export class DataGeneration implements IDataGeneration
@@ -12,7 +12,7 @@ export class DataGeneration implements IDataGeneration
     setStrategy(strategy: string | IDataGenerationStrategy)
     {
         this._strategy = typeof strategy === 'string' ?
-            new (DataGenerationStrategyRegistry.getImplementationByName(strategy)) : strategy;
+            new (dataGenerationStrategyRegistry.getImplementationByName(strategy)) : strategy;
     }
 
     getRentEntries = () => this._strategy.getRentEntries();
