+++ WebScrapingWorker/src/Modules/DataGeneration/IDataGenerationStrategy.ts
@@ -1,37 +1,10 @@
 import { RentEntry } from '../../Models/RentEntry.js';
+import { StrategyRegistry } from '../../Utilities/StrategyRegistry.js';
 
 export interface IDataGenerationStrategy
 {
     getRentEntries(): RentEntry[];
 }
 
-export class DataGenerationStrategyRegistry
-{ // XXX: create an abstract class? extract it to a separate file?
-    private static implementations: (new () => IDataGenerationStrategy)[] = [];
-
-    public static registerImplementation = <T extends new () => IDataGenerationStrategy>(
-        constructor: T
-    ) => { this.implementations.push(constructor); };
-
-    public static getImplementations = () => this.implementations.slice();
-
-    public static getImplementationByName(name: string)
-    {
-        const strategy = this.implementations.find(implementation => implementation.name === name);
-        if (!strategy) throw new StrategyNotFoundError(name);
-        return strategy;
-    }
-}
-
-export class StrategyNotFoundError extends Error
-{
-    constructor(strategyName: string, options: ErrorOptions = {})
-    {
-        const message = `Strategy with name '${strategyName}' not found`;
-
-        super(message, options);
-
-        this.name = this.constructor.name;
-        Error.captureStackTrace?.(this, this.constructor);
-    }
-}
+class DataGenerationStrategyRegistry extends StrategyRegistry<IDataGenerationStrategy> {}
+export const dataGenerationStrategyRegistry = new DataGenerationStrategyRegistry;
