+++ WebScrapingWorker/src/Modules/DataGeneration/Strategies/WebScrapingStrategy.ts
@@ -1,7 +1,7 @@
-import { IDataGenerationStrategy, DataGenerationStrategyRegistry } from '../IDataGenerationStrategy.js';
+import { IDataGenerationStrategy, dataGenerationStrategyRegistry } from '../IDataGenerationStrategy.js';
 import { RentEntry } from '../../../Models/RentEntry.js';
 
-@DataGenerationStrategyRegistry.registerImplementation
+@dataGenerationStrategyRegistry.registerImplementation
 export class WebScrapingStrategy implements IDataGenerationStrategy
 {
     public getRentEntries(): RentEntry[]
